Andrat int till double i addition metoden

+++ src/calculator_alm/Calculate.java
@@ -2,8 +2,8 @@ package calculator_alm;
 
 public class Calculate {
     
-    public int add(int a, int b){
-        int c = a + b;
+    public double add(double a, double b){
+        double c = a + b;
         return c;
     }
